Split out UnaryWhereParameter into multiple files.
