+++ source/Nevermore/ArrayWhereParameter.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using Nevermore.AST;
+
+namespace Nevermore
+{
+    public class ArrayWhereParameter
+    {
+        readonly IReadOnlyList<Parameter> parameterNames;
+
+        public ArrayWhereParameter(string fieldName, ArraySqlOperand operand, IReadOnlyList<Parameter> parameterNames)
+        {
+            this.parameterNames = parameterNames;
+            FieldName = fieldName;
+            Operand = operand;
+        }
+
+        public string FieldName { get; }
+        public ArraySqlOperand Operand { get; }
+        public IEnumerable<string> ParameterNames => parameterNames.Select(p => p.ParameterName);
+    }
+}
